Create htmlgen.pyw
This is the file used to semi-speed up html production, providing a smooth and simple output of individual files. 

Simply input information into fields and hit the generate html button, and it will provide a single html.

This isn't ideal as it it overwrites itself and can only be used for one at a time in it's current state, but you are more than welcome to update that yourself, if I don't do it before you that is.